+++ src/main/java/ru/yandex/practicum/filmorate/storage/GenreStorage.java
@@ -6,6 +6,6 @@ import java.util.List;
 import java.util.Optional;
 
 public interface GenreStorage {
-public Optional<Genre> findGenreById(Integer id);
+public Optional<Genre> findGenreById(Long id);
 public List<Optional<Genre>> findAllGenres();
 }
