+++ src/main/java/ru/yandex/practicum/filmorate/storage/MpaStorage.java
@@ -6,6 +6,6 @@ import java.util.List;
 import java.util.Optional;
 
 public interface MpaStorage {
-    public Optional<Mpa> findMpaById(Integer id);
+    public Optional<Mpa> findMpaById(Long id);
     public List<Optional<Mpa>> findAllMpa();
 }
