+++ src/main/java/ru/yandex/practicum/filmorate/storage/db/UserDbStorage.java
@@ -29,7 +29,7 @@ public class UserDbStorage implements UserStorage {
     }
 
     /*@Override
-    public Map<Integer, Optional<User>> getStorage() {
+    public Map<Long, Optional<User>> getStorage() {
         return null;
     }*/
 
@@ -40,7 +40,7 @@ public class UserDbStorage implements UserStorage {
                 , user.getEmail(), user.getLogin(), user.getName(), Date.valueOf(user.getBirthday()));
         SqlRowSet userRows = jdbcTemplate.queryForRowSet("select USER_ID from USERS where EMAIL=?", user.getEmail());
         if (userRows.next()) {
-            user.setId(userRows.getInt("USER_ID"));
+            user.setId(userRows.getLong("USER_ID"));
         }
         return user;
 
@@ -49,11 +49,11 @@ public class UserDbStorage implements UserStorage {
     //Обновление пользователя
     @Override
     public User update(User user) {
-        Integer id = user.getId();
+        Long id = user.getId();
         SqlRowSet userRows = jdbcTemplate.queryForRowSet("select USER_ID from USERS where USER_ID=?", id);
-        Integer updateUserId = 0;
+        Long updateUserId = 0L;
         if (userRows.next()) {
-            updateUserId = userRows.getInt("USER_ID");
+            updateUserId = userRows.getLong("USER_ID");
         }
         if (updateUserId == id) {
             jdbcTemplate.update("update USERS set EMAIL=?, LOGIN=?, NAME=?, BIRTHDAY=? where USER_ID=?",
@@ -67,11 +67,11 @@ public class UserDbStorage implements UserStorage {
 
     //Получение пользователя
     @Override
-    public Optional<User> getUser(Integer id) {
+    public Optional<User> getUser(Long id) {
         SqlRowSet userRows = jdbcTemplate.queryForRowSet("select * from USERS where USER_ID=?", id);
         if (userRows.next()) {
             User user = new User(
-                    userRows.getInt("USER_ID"),
+                    userRows.getLong("USER_ID"),
                     userRows.getString("EMAIL"),
                     userRows.getString("LOGIN"),
                     userRows.getString("NAME"),
@@ -89,7 +89,7 @@ public class UserDbStorage implements UserStorage {
     @Override
     public List<Optional<User>> getUsers() {
         String sql = "select USER_ID from USERS";
-        List<Optional<User>> users = jdbcTemplate.query(sql, (rs, rowNum) -> getUser(rs.getInt("USER_ID")));
+        List<Optional<User>> users = jdbcTemplate.query(sql, (rs, rowNum) -> getUser(rs.getLong("USER_ID")));
         if (users.isEmpty()) {
             return null;
         } else {
@@ -99,24 +99,24 @@ public class UserDbStorage implements UserStorage {
 
     //Поиск всех друзей пользователя
     @Override
-    public List<Optional<User>> findAllUserFriends(Integer userId) {
+    public List<Optional<User>> findAllUserFriends(Long userId) {
         String sql = "select USER_FRIEND_ID from FRIENDS where USER_ID_PK=" + userId;
-        List<Optional<User>> users = jdbcTemplate.query(sql, (rs, rowNum) -> getUser(rs.getInt("USER_FRIEND_ID")));
+        List<Optional<User>> users = jdbcTemplate.query(sql, (rs, rowNum) -> getUser(rs.getLong("USER_FRIEND_ID")));
         return users;
     }
 
     //список друзей, общих с другим пользователем
     @Override
-    public List<Optional<User>> commonFriends(Integer id, Integer otherId) {
+    public List<Optional<User>> commonFriends(Long id, Long otherId) {
         String userFriends = "select USER_FRIEND_ID from FRIENDS WHERE USER_ID_PK=" + id;
-        List<Integer> userFriendsAll = jdbcTemplate.query(userFriends, (rs, rowNum) ->
-                rs.getInt("USER_FRIEND_ID"));
+        List<Long> userFriendsAll = jdbcTemplate.query(userFriends, (rs, rowNum) ->
+                rs.getLong("USER_FRIEND_ID"));
         String friendFriends = "select USER_FRIEND_ID from FRIENDS WHERE USER_ID_PK=" + otherId;
-        List<Integer> friendFriendsAll = jdbcTemplate.query(friendFriends, (rs, rowNum) ->
-                rs.getInt("USER_FRIEND_ID"));
+        List<Long> friendFriendsAll = jdbcTemplate.query(friendFriends, (rs, rowNum) ->
+                rs.getLong("USER_FRIEND_ID"));
         userFriendsAll.retainAll(friendFriendsAll);
         List<Optional<User>> commonFriends = new ArrayList<>();
-        for (Integer ids : userFriendsAll) {
+        for (Long ids : userFriendsAll) {
             commonFriends.add(getUser(ids));
         }
         return commonFriends;
@@ -124,18 +124,18 @@ public class UserDbStorage implements UserStorage {
 
     //Добавление друга
     @Override
-    public void addFriends(Integer id, Integer friendId) {
-        Integer userIdInBase = 0;
-        Integer friendsIdInBase = 0;
+    public void addFriends(Long id, Long friendId) {
+        Long userIdInBase = 0L;
+        Long friendsIdInBase = 0L;
         SqlRowSet userRows = jdbcTemplate.queryForRowSet(
                 "select USER_ID from USERS where USER_ID=?", id);
         if (userRows.next()) {
-            userIdInBase = (userRows.getInt("USER_ID"));
+            userIdInBase = (userRows.getLong("USER_ID"));
         }
         SqlRowSet userFriendsRows = jdbcTemplate.queryForRowSet(
                 "select USER_ID from USERS where USER_ID=?", friendId);
         if (userFriendsRows.next()) {
-            friendsIdInBase = (userFriendsRows.getInt("USER_ID"));
+            friendsIdInBase = (userFriendsRows.getLong("USER_ID"));
         }
 
         if (userIdInBase > 0 && friendsIdInBase > 0) {
@@ -150,12 +150,12 @@ public class UserDbStorage implements UserStorage {
 
     //Удаление друга
     @Override
-    public void deleteFriend(Integer id, Integer fid) {
-        Integer friendsIdDelete = 0;
+    public void deleteFriend(Long id, Long fid) {
+        Long friendsIdDelete = 0L;
         SqlRowSet userRows = jdbcTemplate.queryForRowSet(
                 "SELECT FRIENDS_ID FROM FRIENDS WHERE USER_ID_PK=? AND USER_FRIEND_ID=?", id, fid);
         if (userRows.next()) {
-            friendsIdDelete = (userRows.getInt("FRIENDS_ID"));
+            friendsIdDelete = (userRows.getLong("FRIENDS_ID"));
         }
         if (friendsIdDelete > 0) {
             jdbcTemplate.update("delete FROM FRIENDS WHERE FRIENDS_ID=?"
