+++ src/main/java/ru/yandex/practicum/filmorate/storage/inmemory/InMemoryFilmStorage.java
@@ -14,11 +14,11 @@ import java.util.stream.Collectors;
 @Slf4j
 @Component("inMemoryFilmStorage")
 public class InMemoryFilmStorage implements FilmStorage {
-    private Map<Integer, Optional<Film>> storage = new HashMap<>();
-    private Integer identifier = 0;
+    private Map<Long, Optional<Film>> storage = new HashMap<>();
+    private Long identifier = 0L;
 
     //Генерация ID номера
-    public Integer generationId() {
+    public Long generationId() {
         identifier += 1;
         return identifier;
     }
@@ -51,7 +51,7 @@ public class InMemoryFilmStorage implements FilmStorage {
     }
 
     //Получение фильма
-    public Optional<Film> getFilm(Integer id) {
+    public Optional<Film> getFilm(Long id) {
         if (storage.containsKey(id)) {
             log.info("Фильм получен.");
             return storage.get(id);
@@ -68,7 +68,7 @@ public class InMemoryFilmStorage implements FilmStorage {
     }
 
     @Override
-    public void addLike(Integer id, Integer userId) {
+    public void addLike(Long id, Long userId) {
         if (storage.containsKey(id) && storage.containsKey(userId)) {
             storage.get(id).get().getLikes().add(userId);
             log.info("Лайк фильму поставлен.");
@@ -79,10 +79,10 @@ public class InMemoryFilmStorage implements FilmStorage {
     }
 
     @Override
-    public List<Optional<Film>> getPopularFilm(Integer count) {
+    public List<Optional<Film>> getPopularFilm(Long count) {
         log.info("Получен список отсотритованных фильмов.");
         if (count <= 0) {
-            count = 10;
+            count = 10L;
         }
 
         return storage.values().stream()
@@ -92,7 +92,7 @@ public class InMemoryFilmStorage implements FilmStorage {
     }
 
     @Override
-    public void deleteLike(Integer id, Integer userId) {
+    public void deleteLike(Long id, Long userId) {
         if (storage.containsKey(id) && storage.get(id).get().getLikes().contains(userId)) {
             storage.get(id).get().getLikes().remove(userId);
             log.info("Лайк у фильма  удален.");
